Db -> Analysis dry run

+++ src/engagement_db_to_analysis/engagement_db_to_analysis.py
@@ -50,7 +50,7 @@ def export_traced_data(traced_data, export_path):
 
 
 def generate_analysis_files(user, google_cloud_credentials_file_path, pipeline_config, engagement_db, membership_group_dir_path,
-                            output_dir, cache_path=None):
+                            output_dir, cache_path=None, dry_run=False):
     analysis_dataset_configurations = pipeline_config.analysis.dataset_configurations
     # TODO: Tidy up which functions get passed analysis_configs and which get passed dataset_configurations
 
@@ -64,7 +64,7 @@ def generate_analysis_files(user, google_cloud_credentials_file_path, pipeline_c
     engagement_db_datasets = []
     for config in analysis_dataset_configurations:
         engagement_db_datasets.extend(config.engagement_db_datasets)
-    messages_map = get_messages_in_datasets(engagement_db, engagement_db_datasets, cache)
+    messages_map = get_messages_in_datasets(engagement_db, engagement_db_datasets, cache, dry_run)
 
     messages_traced_data = _convert_messages_to_traced_data(user, messages_map)
 
@@ -108,19 +108,23 @@ def generate_analysis_files(user, google_cloud_credentials_file_path, pipeline_c
 
     run_automated_analysis(messages_by_column, participants_by_column, pipeline_config.analysis, f"{output_dir}/automated-analysis")
 
+    dry_run_text = "(dry run)" if dry_run else ""
     if pipeline_config.analysis.google_drive_upload is None:
-        log.debug("Not uploading to Google Drive, because the 'google_drive_upload' configuration was None")
+        log.debug(f"Not uploading to Google Drive, because the 'google_drive_upload' configuration was None {dry_run_text}")
     else:
-        log.info("Uploading outputs to Google Drive...")
-        google_drive_upload.init_client(
-            google_cloud_credentials_file_path,
-            pipeline_config.analysis.google_drive_upload.credentials_file_url
-        )
+        if dry_run:
+            log.info(f"Not uploading to Google Drive {dry_run_text}")
+        else:
+            log.info("Uploading outputs to Google Drive...")
+            google_drive_upload.init_client(
+                google_cloud_credentials_file_path,
+                pipeline_config.analysis.google_drive_upload.credentials_file_url
+            )
 
-        drive_dir = pipeline_config.analysis.google_drive_upload.drive_dir
-        google_drive_upload.upload_file(f"{output_dir}/production.csv", drive_dir)
-        google_drive_upload.upload_file(f"{output_dir}/messages.csv", drive_dir)
-        google_drive_upload.upload_file(f"{output_dir}/participants.csv", drive_dir)
-        google_drive_upload.upload_all_files_in_dir(
-            f"{output_dir}/automated-analysis", f"{drive_dir}/automated-analysis", recursive=True
-        )
+            drive_dir = pipeline_config.analysis.google_drive_upload.drive_dir
+            google_drive_upload.upload_file(f"{output_dir}/production.csv", drive_dir)
+            google_drive_upload.upload_file(f"{output_dir}/messages.csv", drive_dir)
+            google_drive_upload.upload_file(f"{output_dir}/participants.csv", drive_dir)
+            google_drive_upload.upload_all_files_in_dir(
+                f"{output_dir}/automated-analysis", f"{drive_dir}/automated-analysis", recursive=True
+            )
